mvc set up
