--- StudyGroup/StudyGroup/Student.cs
@@ -1,13 +1,0 @@
-﻿namespace StudyGroup;
-public class Student
-{
-    public Student()
-    {
-
-    }
-    public int Id { get; set; }
-    public string Name { get; set; }
-    public string Address { get; set; }
-    public bool AttendanceRecorded { get; set; }
-    public bool AttendanceValue { get; set; }
-}
