--- StudyGroup/StudyGroup/TeachingAssistant.cs
@@ -1,7 +1,0 @@
-﻿namespace StudyGroup;
-public class TeachingAssistant : Student // teaching assistant inherits student properties and access to functions (if it was public) 
-{
-    public string CollegeEmailAddress { get; set; }
-    public bool TookAttendance { get; set; }
-}
-
